Call middle parts of MeowURI "children".
Renames the 'MeowURINode' class to 'MeowURIChild'

Changes other usage of a "node" to "child", "nodes" to "children".

The reason is that a "node" probably is a better describes for all
parts of a MeowURI, which might be joined in an abstract base class
with this name.
This is also more similar to basic computer science terminology.