Add new 'ConfigurationRuleParser' class.
Adds new class to replace parts of the 'ConfigurationParser'.

Adds new TODO [TD0139] on warning if a rule lists data sources
for placeholder fields but do not use the fields in the actual
name template. This could happen if a "shared" name template is
misspelled, which currently results in it being used as a name
template, without any placeholder fields.

That is, the user intended it to be a reference to an entry in
'NAME_TEMPLATES' but it is instead used as a "literal" without
any warnings or indications.