Rename default test configuration file.
Renames the default config used by various tests and fixes access.

Renames 'test_files/default_config.yaml' to
'test_files/configs/default.yaml'.

Fixes various ways in which this path is accessed so that making
changes similar to this again in the future won't be as messy.

Mainly, adds a new unit test utility function 'abspath_testconfig()'
that should be used to get the absolute path to the default config.