Add limiting size of cache files.
Implements ability to limit the size of cached files.

Limits are checked every time data is written to the cache, which
probably should be reworked at some future point. It would probably
make more sense to check the cache size at instantiation or possibly
when the cache leaves the scope, using the context-manager protocol.

When the cache file size exceeds the limit, the n oldest items are
removed from both RAM and disk. The number n is currently half the
number of items, so if the cache contains 100 items, the 50 oldest
are removed when the cache file size exceeds the limit.

This completes TODO-list item [TD0101].