Pass '-OO' to interpreter in runner script.
Adds using Python interpreter option '-OO' in 'bin/autonameow.sh'.

From 'man python3':

    -O    Turn on basic optimizations.
          Given twice, causes docstrings to be discarded.
    -OO   Discard docstrings in addition to the -O optimizations.

This change is mostly experimental. Basic measurements show that actual
performance improvements are marginal at best. While some use-cases run
10-20% faster with the optimizations enabled, it is still not clear if
this is worth sacrificing assertions and debugging facilities.
If one actually was to try to improve the performance, this is not the
way to go about it.

The main motivation is rather experimenting with conditional execution of
(known from profiling) expensive operations in "if __debug__:" conditions.