Fix wrong key used in 'maps_field()'.
Fixes bug in 'maps_field()' due to using the wrong key in dict lookup.

This is a leftover from when removing the 'ExtractedData' class that
had a property (or function (?)) named 'field_map'.
This was simply translated to a dict lookup using a key with the same
name, but the key actually used by providers was 'mapped_fields', so
accesses would always fail.

The function 'maps_field()' would always return False due to this bug.