Refactor getting candidates from the resolver.
Refactors queries to the resolver for field candidates.

Moves converting the candidates (data dicts) to string representations
to the resolver class. This also adds returning the values wrapped in a
'FieldDataCandidate' class. This is currently only used by functions in
the interactive module.

This and many related things are **VERY HACKY** and should be reworked.

At least for now, this prevents crashing.