Remove module attribute 'MEOWURI_SOURCE_MAP_DICT'.
Removes global module attribute holding MeowURI-class-mappings.

This was a leftover from before the 'ProviderRegistry' class which
now provides an interface to access this information.

The nested function and global variable was meant to prevent repeated
imports of the 'analyzers', 'extractors' and 'plugins' modules.
But this is no longer necessary and could in fact be a source of bugs;
if one were to "re-initialize" the repository by calling doing so might
well be to detect changes in available providers, which the now removed
solution would actually prevent.