Inline modified 'run_task()' in integration runner.
Removes use of and sourcing 'common_utils.sh' from integration runner.

Inlines a modified 'run_task()' that does not attempt to handle
aborting the running task, because it did not work.

This also sources the test suite scripts instead of evaluating
('eval x') them. This change is part of adding a global counter and
attempting to solve shared state between all integration test scripts
in the mess of sourcing/"evaling".