Add translating strings to generic field classes.
Modifies specifying generic fields in provider 'FIELD_LOOKUP'.

Instead of listing the generic field subclasses directly, which
requires importing the module; specify the generic field classes
as simple strings and "translate" these to classes in the
respective provider "runner".

This might be more robust to future changes and also allows
further experimentation with reworking generic fields and
integrating the concept of records.