Modify handling of "generic fields" in analyzers.
Adds fixes to handling of "generic fields" in analyzers.

The generic fields were accessed in a try/catch 'AttributeError'
clause in the repository as to not crash completely when random,
inconsistent data was passed in.

Hack fixes in '_add_results()', used by analyzer classes to store
data in the repository allowed removing the try/catch in the
repository.  However, this is still in great need of improvements
and possibly a complete rewrite.

Adds TODO [TD0146] on reworking "generic fields" and possibly
bundling multiple fields in some type of "records".