Factor out '_handle_file()' into new class.
Moves contents of 'Autonameow._handle_file()' to a new class.

Adds a new class 'FilesContext' in 'core/context.py'.
The contents of 'Autonameow._handle_file()' is moved to
'FilesContext.handle_file()' without substantial functional
modifications.

This is motivated by future changes, where files within the same
"context" are handled together in a batch. It also mitigates the
growing complexity and length of code in the 'Autonameow' class.