Add (WIP) rework of overall architecture.
Adds initial rework of "data pipeline" as per [TD0142].

NOTE:  This commit passes tests but only because the OCR extractor
       is not covered by regression tests! CONSIDER BROKEN AS-IS.

A new subsystem is added to handle top-most level data requests.
The 'MasterDataProvider' class is not completed. It is intended to
take data requests passed as a FileObject instance and a MeowURI.
(I.E. data for a certain file, matching a certain identifier)

If the data is not stored in the repository, the appropriate provider
(extractor/analyzer/plugin) should be tasked with gathering the data.
If successful, the requested data is stored in the repository.
Store any other data, gathered as a corollary as well.
E.G. if running 'exiftool' to get a specific field, DO NOT discard
other fields, keep them to save on calls..

If the data IS stored in the repository, return it directly.