Rework regression test monkey-patching.
Fixes regression test monkey-patching of the rename method.

Previously, even files that were skipped because the new name was equal
to the old name was counted as renamed files.
These are listed in the regression tests 'asserts.yaml'.

This commit reworks this to at least solve some of the issues, but
also causes breakage of existing regression tests that now
produce false negatives.
Changing the regression tests to not count a skipped file as a rename
makes much more sense anyway, so it is a change in the right direction.