Do not pass available extractors to runner.
Removes argument 'available_extractors' from 'ExtractorRunner' init.

This class currently only gets accessed through the module-level
'run_extraction()' function, which populated this argument.  But this
is pretty much equivalent to having the class grab this data itself at
instantiation.

The old method might provide a means of controlling which extractors
should be made available to the runner, but for now this probably only
adds unnecessary complexity.