Improve robustness of 'MeowURIParser.parse()'.
Adds fixes and improved robustness to 'MeowURIParser.parse()'.

Structures of lists and/or tuples nested deeper than one level was
not handled properly. It now should be able to accept just about
any combination of arguments, like for instance;

    mp.parse('a')
    mp.parse('a', 'b')
    mp.parse(['a', 'b'])
    mp.parse(['a', 'b'], 'c')
    mp.parse(['a', 'b'], 'c', ['d'])
    mp.parse(['a', 'b'], 'c', (['d'], 'e', ['f', ['g', 'h']]))

The alternate solution would be to restrict how this method is being
called to some expected usage, which might actually make more sense,
but since this is intended to be a user-facing "parser", this extra
ruggedness is not _entirely_ misplaced.