Fix checking tty before running 'prompt_toolkit'.
Adds checking that stdin is a tty. Previously checked stdout.

Even though previous check probably worked most of the time it
is better to reflect the assertions in 'prompt_toolkit/input.py';

    class StdinInput(Input):
        """
        Simple wrapper around stdin.
        """
        def __init__(self, stdin=None):
            self.stdin = stdin or sys.stdin

            # The input object should be a TTY.
            assert self.stdin.isatty()

One check is also moved so they are all contained in 'prompt.py'
which works as a wrapper around 'prompt_toolkit'.