Fix regressions introduced in 353fcd74100d68b6523.
Fixes regressions introduced when fixing skipping some metadata.

Changes in commit 353fcd74100d68b6523f17ee7c276c7c8c221dc2 caused
some of the regression tests that used generic data sources for
name template fields.
The changes caused additional valid metadata to be extracted,
causing the resolver to fail because it could not pick out which
of multiple possible candidates to use.

This is now fixed by adding a simple means of picking out one piece
of data from a list of candidates, based on the hardcoded
'WeightedMapping' probabilities that is assigned by providers.

So given a "generic" MeowURI such as 'generic.metadata.title',
the candidate that has the highest WeightedMapping probability
to the name template field 'Title' is used.
The current solution will have to be reworked to work in a general
case and be less confusing ..