Remove 'FiletagsAnalyzer'. Add 'FiletagsExtractor'.
Reworks the filetags analyzer into a extractor.

Functionality should be basically unchanged.
Because the 'FiletagsAnalyzer' was very simple and did not require much
of "contextual information" it is arguably better suited as a extractor.

This was also in part an exercise in evaluating the overall architecture
and how much friction is involved in creating new extractors.

MeowURIs 'analyzer.filetags.*' are now 'extractor.filesystem.filetags.*'