Do not assume at least one analyzer applies.
This no longer holds after disabling and reworking analyzers.

The 'ImageAnalyzer' and 'TextAnalyzer' was disabled because they do not
actually do and return results in a format that deviates from the other
providers.
And the 'FiletagsAnalyzer' was rework to a extractor.

It could now very well be that no analyzers applies to a given file.