Remove searching for datetime in 'DocumentAnalyzer'.
Removes passing back date/time-information using older method.

This inline wrapping of results dicts in a list does not work with
the other systems that use 'FIELD_LOOKUP' to wrap values into the
dicts.  The returned 'value' would be a list of additional dicts,
which would break systems that receive this data.