Rename 'PdftotextTextExtractor' to 'PdfTextExtractor'.
Renames the pdf text extractor class and related source files.

Makes a lot more sense to not bother with leaving room for many
different extractors for the same file type. Users probably
will not be interested in specifying exactly which back-end program
is used to get the text.

Users should not even have to specify which extractor class to use,
something like 'extractor.text.full' should be enough.