Rework extractor unit test inheritance.
Improves use base classes in extractor unit tests.

Previously a lot of tests were skipped, printing warnings and
polluting the results. This was caused by having the base classes
with shared functionality inherit from 'unittest.TestCase'.
This commit fixes this by only inheriting from 'unittest.TestCase'
in the test classes that use the shared functionality.

The same change is also made to the property-based tests of the
config field parsers.