Add dependencies to 'setup.py'.
Adds dependencies with the lowest versions currently used for testing.

I test autonameow regularly on MacOS and Linux and the versions of
python modules available through homebrew are all newer than those
found in the Ubuntu repositories.  The versions listed here are the
ones used on my Linux machine.  Full listing, for future reference:

                        Linux       MacOS
    beautifulsoup4      4.4.1       4.6.0
    chardet             2.3.0       3.0.4
    colorama         *  0.3.7       0.3.7
    EbookLib         *  0.16        0.16
    hypothesis       *  3.38.0      3.38.0
    guessit             2.1.0       2.1.4
    isbnlib             3.5.6       3.7.2
    Pillow              3.1.2       4.0.0
    pytz                2014.10     2017.2
    PyYAML              3.11        3.12
    Unidecode           0.4.19      0.4.20

Note that "magic" is missing because I can't seem to be certain
of exactly which library is actually used on both platforms ..