Prevent '(UNKNOWN MIME-TYPE)' to be used as extension.
Adds testing if an instance of 'NullMIMEType' is used as extension.

Checks if the value is non-empty in the 'Extension' name template
field 'format()' method.
However, this test should probably be done a lot earlier, as there
is a distinct difference between an unknown MIME-type and a MIME-type
that is mapped to an empty extension.