Add parser for external 'FilenameAnalyzer' data.
Adds a 'MimetypeExtensionMapParser' class for reading data file.

The function 'likely_extension()' uses data currently stored in dict
'MIMETYPE_EXTENSION_SUFFIXES_MAP'.
This data should be stored in an external file.
I first tried dumping the data in YAML format, but the aliasing and
handling of sets, etc.  gave a pretty messy result. This commit adds
a way overkill, poorly implemented parser for reading the simple data
file.

Note that this commit does not actually swap out the old dict with
this new parser and data.

The file 'probable_extension_lookup' is also added to the manifest
and 'setup.py'.