Add external 'FIELD_LOOKUP' files and functions.
Adds functionality for reading 'FIELD_LOOKUP' from yaml files.

Also adds the yaml-files themselves.

Note that the previous way of defining 'FIELD_LOOKUP' as class
instance attributes is still in use.  This commit just adds the
functionality and data files that contain the same information
as that in the 'FIELD_LOOKUP' dicts.