Increase size of text chunks for epub e-books.
Increases the amount of text that is searched for ISBN-numbers.

The 'EpubTextExtractor' returns text with a lot of duplicated lines and
sections. I'm currently not sure if this is to be expected or caused by
the "hail mary" approach currently used in the 'EpubTextExtractor',
collecting all possible text ..

Assuming that this is typical for epub e-books, the text chunks that are
searched for ISBN-numbers are increased for files with MIME-type
'application/epub+zip'.