Remove blacklisted text lines in text extractors.
Adds filtering blacklisted text lines in the 'AbstractTextExtractor'.

Part of the 'cleanup()' method that is called with the raw extracted
in all text extractors.
This should probably allow regular expressions, but this simple exact
matching is a start and does not risk increasing extractor run-time
as much as regular expressions *probably* would.