Add workaround to getting provider names.
Adds workaround for being unable to get name from 'str(klass)'.

Adds class method 'name()' to 'BaseAnalyzer' and 'BaseExtractor'.

Should pass around instances of provider classes instead of the
classes themselves to solve this and other related problems.
This has not been done yet because that kind of refactoring would
likely introduce a lot of problems related to caching in the text
extractors and ebook analyzer ..

References TODO-list item [TD0151].