Factor out 'isbnlib' functionality to new module.
Adds a new module 'services' in the source root.

This module should contain third-party services for querying APIs.

Most of the 'isbnlib' functionality is moved from 'analyze_ebook.py'
to a new file 'services/isbn.py'.

This should be cleaned up and designed as to provide an interface
to *any* ISBN metadata service provider, not just 'isbnlib'.
The interface should also handle take care of verifying that the
services are available and other service-specific details such as
rate-limiting, etc.

References TODO-list item [TD0182].