Move 'FilenamePostprocessor' to outer scope.
Moves the 'FilenamePostprocessor' from the name builder.

This was the fastest way to bypass the 'TemplateFieldDataResolver'
when the name template of the used rule does not contain any
placeholder fields.

Adds new regression test suites with a single rule with a "static"
name template, I.E. no placeholder fields and also no data sources.