Modify coercing provider result values.
Adds skipping values with unknown (None) 'multivalued'.

Instead of implicitly defaulting to 'multivalued' being False when
unspecified/malformed/None, skip the coercion of the value.
This means that the value will not be included in the final results
stored in the repository.
A lot of problems are related to the confusion around whether values
are lists of single values; a single value which *IS* a list and there
consists of multiple values, or multiple values that each contain
multiple values.
Makes sense to remove as much ambiguity as possible until this is less
of a total mess ..

Also other minor fixes.