Use a single shared 'exiftool' process.
Modifies extractors to be able to re-use a single external process.

References TODO-list item [TD0183].
Still not robust enough to be considered completed.

This now requires all extractors to implement a 'shutdown()' method
which is actually only relevant for the 'ExiftoolMetadataExtractor'.
Other extractors do not keep any such state to "de-initialize".