Skip coercion of missing filetags "timestamps".
Adds checking if "timestamp" is None to prevent failed coercions.

Set timestamp to None instead of empty string here so that it can be
detected and skipped when converting values to the "internal format".
Coercing None with 'AW_TIMEDATE' raises a 'AWTypeError' exception,
which would happen for every file that do not have a "timestamp"
filetags part.