Fix resolver not reset when using new rule.
Fixes bug where the resolver was not properly reset "between rules".

After trying and failing to apply a given rule the resolver state was
kept, which led to attempts to gather data for placeholder fields that
were present in the rule active previously but not present in the
currently active rule.

This part of the code is all just a total mess, I'm surprised it
works as "well" as it does..