Add method for finding "probable" ISBN metadata.
Adds a first attempt at finding most probable ISBN metadata.

This is very hacky and mostly a proof of concept that should
be handled by some separate system.
This will most likely work with the yet unimplemented concept
of "records" storing collection of "fields".

Selection of a "reference" field to use for determining the
"probability" of a (ISBN metadata) record must be more flexible.