Store references to "generic" data in 'Repository'.
Reworks the 'Repository' to store references to "generic" data.

Previous implementation stored all data in duplicates, both under the
"explicit" full MeowURI as well as under the "generic" MeowURI, if any.

This commit adds mapping generic URIs to explicit URIs. When storing
data with 'generic_field' in the "metainfo", the generic field is added
in a fileobject-specific lookup table.
When handling a query with a "generic" MeowURI, any corresponding
"explicit" URIs is retrieved from the lookup table and used to retrieve
the data.