Fix crashing if mismatched candidate and field.
Fixes crashing due to formatting fields with incompatible value.

Fixes crashing when attempting to format field candidate with a
placeholder field that are not both either "multivalued" or not.

Fixed by checking that the placeholder field is compatible with the
field candidate before calling the 'format()' method which raised a
assertion error exception.