Rework string value canonicalization.
Various rework related to the 'StringValueCanonicalizer'.

Modifies the format of the config yaml file.
Patterns are now either literal strings or regular expressions.

Example yaml config section:

    CanonicalValue:
        match_any_regex:
        - regex pattern 1
        - regex pattern 2
        match_any_literal:
        - string pattern 1
        - string pattern 2

Any value that matches either 'CanonicalValue', any of the
regular expressions or any of the literal strings is replaced
with 'CanonicalValue'.

The "parsing" of the regular expressions and literals is moved
to a separate class 'CanonicalizerConfigParser' so that the
'StringValueCanonicalizer' class can be passed plain dicts keyed
by canonical values, storing lists of either literal strings or
compiled regular expressions.

This also adds handling of matching multiple regular expressions,
in which case the canonical value whose regular expression produced
the longest matched substring is used.