Fix repeated provider delegation bug.
Fixes repeated delegation of providers when running all providers.

The 'ProviderRunner' checks if a provider has already been delegated
and skips the provider to prevent unnecessarily running extractors or
analyzers more than once times and subsequently storing duplicate data.

This check did not account for calls to the runner to delegate to all
possible providers through 'delegate_every_possible_meowuri()', which
triggered an assertion;

    AssertionError: Would have clobbered value for URI
    extractor.metadata.exiftool.File:FileSize

This is now fixed by keeping track of the files that are passed to
'delegate_every_possible_meowuri()' and skipping any delegations with
those files.