Fix 'MeowURIMapper' not storing all explicit URIs.
Fixes bug that led to only storing the most recently mapped URI.

The mapping of a "generic" (or "aliased") URI leaf is not symmetrical,
there will almost always be more than one full "explicit" URI mapped
to one URI with a "generic" leaf.

These are now stored in a set and the value returned by calling
'fetch()' will always be a set, even if it only contains a single
element. This should be sorted out somewhere else (repository) ..