Fix failed assertion in resolver.
Fixes triggering assertion error in the resolver.

This happened when the response contained multiple entries but the
URI was not "generic", which could not happen previously when the
"leaf aliases" was still not implemented.

Now however, a URI with a "generic" leaf could expand to multiple
explicit URIs and return multiple results, which was not handle by
the mess of if-else-statements in the resolver.

This is now "fixed" by aborting when this happens ..

    NOTE: This should really be fixed properly!
          See TODO-list item [TD0112]