Decorate 'git_commit_hash()' with 'lru_cache'.
Wraps 'git_commit_hash()' with the 'lru_cache' decorator.

This should work even if a running autonameow process is left
running for a long time and the currently checked out code changes.
The running code should not be reloaded from disk, so the only
thing that cause confusion is if files other than Python sources;
such as yaml configuration files, etc., are changed during runtime.