Improve results of 'find_publisher()'.
Adds also searching the "canonical" forms of publishers.

The 'FilenameAnalyzer' uses this "old" method of using
"known good" field values specified in the user configuration
to find publishers in unstructured text.

This should be replaced with something related to the newer
systems related to 'autonameow/core/metadata/canonicalize.py'.
Something new probably needs to be added to separate the fuzzy
searching patterns from the canonical replacement patterns as
they likely will have to be different for optimal results.