Fix skipping some templates without placeholders.
Fixes a bug in the horrible code in 'context.py'.

If there are no data sources available but the name template does not
include any placeholder fields, the name template can simply be used
as-is, without having to modify it further.

This was only checked for the first evaluated rule, which meant that
when additional rules are tried ("automagic mode") only the first rule
was properly handled.  When any following rules were tried, they were
skipped because the data sources was wrongly assumed required and also
not specified in that rule.

Also adds regression tests that would have triggered this bug.