Remove 'RepositoryPool'.
Removes 'RepositoryPool' and modifies repository (de-)initialization.

The original intent was to always keep one 'Repository' instance linked
to one running instance, represented as an instance of the 'Autonameow'
class.

This worked in some respects, but the manner in which the module-level
attribute 'SessionRepository' was accessed, this extra complexity
actually provided no real benefits as the access did not take either
multiple parallell repositories or autonameow instances into account.

The 'SessionRepository' is now a simple global that is instantiated,
"initialized" and finally "de-initialized" through the 'on_startup'
and 'on_shutdown' event handlers, respectively.

Note that this always makes the 'autonameow_instance' passed to the
event handlers completely pointless apart from the debug logging.