Add ability to set 'ColumnFormatter' width limit.
Adds ability to set a maximum total width of formatted output.

If 'max_total_width' is set before calling 'str()' on the
'ColumnFormatter' instance, the width of fields will be truncated
until the total width is less than the set 'max_total_width'.

This is not used by default, so the default behaviour is unchanged.