Catch all exceptions whens calling 'ebooklib'.
Handle all exceptions to ensure not crashing due to 'ebooklib' bugs.

Various AttributeError and KeyError exceptions are raised for some
epub files. Happens infrequently but results in a hard crash if not
properly handled.
Instead of listing all possible uncaught exceptions that is not not
caught and re-raised as 'EpubException' by 'ebooklib', it is more
practical and safer to simply catch all exceptions.