Improve 'FiletagsExtractor' handling of dates.
Adds handling "filetags file names" containing dates without time.

Various fixes to the 'FiletagsExtractor' to highlight the fact that a
valid filetags "timestamp" can be either just a date (~"YYYY-mm-dd")
or both date and time (~"YYYY-mm-ddTHHMMSS").

Yet again the choices made when designing the system with "field meta"
it too rigid and inflexible. In this case, when a timestamp can be
either a date or a datetime, there is no clean way to declare this in
the YAML meta info file and no non-hacky ways to workaround the type
coercion system.
So, currently either none or either of 'date' and 'datetime' will be
used depending on whether the timestamp exists and if it contains a
date or both date and time __BUT__ this will raise an error which will
be debug logged most of the time, even though it is not really an
error..