Fix 'split_basename_prefix_into_filetags_parts()'.
Modifies 'split_basename_prefix_into_filetags_parts()'.

Various refactoring for readability as well as changes to the
behaviour of function used by the 'FiletagsExtractor' to split
a basename without extension into separate "filetags parts".

This now always returns the description as a Unicode string.
When there is no "filetags description" present, the description
is returned as an empty string.

Adds additional test cases that verify this improved behaviour.