Refer to a rule "weight" as its "relative score".
Rename 'weight' to 'relative_score' in code.

The score of a rule is the ratio of conditions met to the number
of conditions.
The "relative score" is the ratio of conditions met to the number
of conditions in the rule with the _maximum_ number of conditions,
among the rules that was evaluated at the same time.