Rework basename splitting function return values.
Replaces returning None values for missing parts.

Return 'EMPTY_FILENAME_PART' instead of None for empty "parts".
Given a basename "foo.bar", the prefix is "foo" and "bar" is the
suffix.  Basenames without either part; like ".foo" or "foo",
are split into tuples of ("foo", EMPTY_FILENAME_PART).

This makes for a cleaner interface and removes the need for callers
to explicitly check for and replace None values.