Fix crash when listing multiple excluded providers.
Fixes possible crash when running '--dump-meowuris --verbose'.

If multiple providers are excluded, the 'sorted()' call fails
with a TypeError due to the providers does not have a common
base class that makes the types unorderable.

Fixed by adding an accessor that sorts the providers by class
names each time the property is accessed.