Assert readable 'vendor' directory exists.
Adds sanity-check assertion before modifying 'sys.path'.

Asserts that the effective UID/GID can list and "cd" to the
'vendor' directory; R_OK for access and X_OK for being able
to enter and write _into_ the directory.
Note that this uses the real UID and GID, not the effective
ones.  This may prove to be a problem down the line ..