Implement querying metadata for ISBN numbers.
Improves finding ISBN numbers in the 'EbookAnalyzer'.

If the textual contents does not turn up a match, the
'generic.metadata.identifier' MeowURI is queried and any
results are run through the same text extraction functions
as used for searching through the textual contents.

This drastically improves the chances of finding valid ISBN
numbers for files of some types and contents.