Add ability to grab all possible CLI options.
Adds function 'get_gnu_style_optionals_from_argparser()'.

Intended to be used primarily (solely) by the regression test
runner; replacing the currently hardcoded list of options in
'tests/regression/utils.py' with a mechanism that fetches the
data by introspecting an instance of 'ArgumentParser'.

Returns *options of interest for the regression runner* in
the longer "GNU-style" form ("--help" as compared to "-h") as
a set of Unicode strings.