Handle unset '$TERM' and/or missing 'tput'.
Adds checking that '$TERM' is set and that 'tput' is executable.

Disable ANSI colors and attributes in bash scripts if '$TERM' is
unset, which could happen when connecting over SSH and within the
integrated PyCharm terminal. Also don't try to use 'tput' before
checking that it is actually installed and executable.