Rewrite 'ArgumentParser' introspection function.
Removes 'get_gnu_style_optionals_from_argparser()'

Adds a new function 'get_optional_argparser_options()'.

Grabs information *options of interest for the regression runner*
by introspecting an instance of 'ArgumentParser'.

Intended to be used primarily (solely) by the regression tests so
that options only need to be defined in this file as compared to
manually keeping changes in sync.
Returns *options of interest for the regression runner* as tuples
of three strings; [SHORT] [GNU] [DEST]:

  SHORT Short form of the command-line option ("-h")
  LONG  Longer "GNU-style" form of the command-line option ("--help")
  DEST  Name of the attribute that stores the option value.