Add ability to get known incorrect field values.
Adds new function 'incorrect_values()' with unit tests.

Returns values known to be incorrect for the given field name.

Accepts a field value that corresponds to the basenames of files
stored in the 'core/truths/data' directory and returns a set of
all values known to be incorrect for the given field name.