Modify how 'Rule.__gt__()' compares instances.
Attempt at fixing inconsistent behaviour of rule listing.

Compare 'Rule' instances by comparing additional attributes.

Also adds more unit tests in the (failed) process of debugging
non-deterministic results of the local regression testsuite '9028'.

However, it _seems_ that the problem actually lies in the function
'prioritize_rules()' in 'rulematcher.py'; that does the sorting
of rules prior to rule evaluation. This is is still not verified.