Move 'BaseMetadataExtractor' to related package.
Moves the 'BaseMetadataExtractor' to the 'metadata' package.

This base class is only ever inherited from by classes in the
'autonameow.extractors.metadata' module and so it makes sense to move
the file containing the base class, 'base.py' to the same package.

This also makes the structure consistent with that of the text extractor
base class 'BaseTextExtractor' and inheriting classes, in the
'autonameow.extractors.text' module.